Fairseq: Save predictions in logging output for evaluating MAP and MAUC (#1499)
Summary:
Pull Request resolved: https://github.com/pytorch/fairseq/pull/1499

In sound event detection, the evaluation metrics are mean average precision (MAP) and mean area under the curve (MAUC).
These metrics evaluate how well the system *sorts* instances.
They are not decomposable across instances, so one needs to collect the predictions and truths of all instances of a validation set.

This diff adds a switch `log_pred` to the `forward` function of the binary cross entropy.
When this switch is turned on, it saves the predictions and truths to the `logging_output` dictionary.

This dictionary needs to be synced across workers by the `all_gather_list` function in `distributed_utils.py`.
The existing implementation restricts the size of the dictionary to 64KB, because it only allocates 2 bytes for the size during serialization. The predictions and truths often exceed this limit.
This diff uses 4 bytes for the size during serialization, and increases the limit.

Reviewed By: myleott

Differential Revision: D18922980

fbshipit-source-id: 1210b85c62dafab729de1fe0eced9a0cc56814a4